safer test init, md5 as parameters instead of file, test xml files names with extension
Former-commit-id: 9ed62ca73dfe419b32d44d73de73244372d3074c

+++ prototype10/src/test/java/cz/inqool/arclib/AipApiTest.java
@@ -11,7 +11,10 @@ import cz.inqool.arclib.storage.StorageType;
 import cz.inqool.arclib.store.AipSipStore;
 import cz.inqool.arclib.store.AipXmlStore;
 import org.apache.commons.io.FileUtils;
-import org.junit.*;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.BeforeClass;
+import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.mock.web.MockMultipartFile;
@@ -55,15 +58,13 @@ public class AipApiTest extends DbTest implements ApiTest {
     private static final String SIP_HASH = "CB30ACE944A440F77D6F99040D2DE1F2";
     private static final Path SIP_PATH = Paths.get("sip", SIP_ID.substring(0, 2), SIP_ID.substring(2, 4), SIP_ID.substring(4, 6));
 
-
     private static final String XML1_ID = "XML1testID";
-    private static final String XML1_FILE_NAME = "xml1";
+    private static final String XML1_FILE_NAME = "xml1.xml";
     private static final String XML1_HASH = "F09E5F27526A0ED7EC5B2D9D5C0B53CF";
     private static final Path XML1_PATH = Paths.get("xml", XML1_ID.substring(0, 2), XML1_ID.substring(2, 4), XML1_ID.substring(4, 6));
 
-
     private static final String XML2_ID = "XML2testID";
-    private static final String XML2_FILE_NAME = "xml2";
+    private static final String XML2_FILE_NAME = "xml2.xml";
     private static final String XML2_HASH = "D5B6402517014CF00C223D6A785A4230";
     private static final Path XML2_PATH = Paths.get("xml", XML2_ID.substring(0, 2), XML2_ID.substring(2, 4), XML2_ID.substring(4, 6));
 
@@ -72,18 +73,16 @@ public class AipApiTest extends DbTest implements ApiTest {
 
     @BeforeClass
     public static void setUp() throws IOException {
+        if (Files.isDirectory(Paths.get("sip")))
+            FileUtils.deleteDirectory(new File("sip"));
+        if (Files.isDirectory(Paths.get("xml")))
+            FileUtils.deleteDirectory(new File("xml"));
         Files.createDirectories(SIP_PATH);
         Files.createDirectories(XML1_PATH);
         Files.createDirectories(XML2_PATH);
         Files.copy(Paths.get(SIP_SOURCE_PATH.toString()), SIP_PATH.resolve(SIP_ID));
         Files.copy(Paths.get("./src/test/resources/aip/" + XML1_FILE_NAME), XML1_PATH.resolve(XML1_ID));
         Files.copy(Paths.get("./src/test/resources/aip/" + XML2_FILE_NAME), XML2_PATH.resolve(XML2_ID));
-    }
-
-    @AfterClass
-    public static void cleanUp() throws IOException {
-        FileUtils.deleteDirectory(new File("sip"));
-        FileUtils.deleteDirectory(new File("xml"));
     }
 
     @Before
@@ -131,11 +130,9 @@ public class AipApiTest extends DbTest implements ApiTest {
                 "sip", "sip", "text/plain", Files.readAllBytes(SIP_SOURCE_PATH));
         MockMultipartFile xmlFile = new MockMultipartFile(
                 "xml", "xml", "text/plain", XML1_ID.getBytes());
-        MockMultipartFile metaFile = new MockMultipartFile(
-                "meta", "meta", "text/plain", (SIP_HASH + System.lineSeparator() + XML1_HASH).getBytes());
 
         mvc(api)
-                .perform(MockMvcRequestBuilders.fileUpload(BASE + "/store").file(sipFile).file(xmlFile).file(metaFile))
+                .perform(MockMvcRequestBuilders.fileUpload(BASE + "/store").file(sipFile).file(xmlFile).param("sipMD5", SIP_HASH).param("xmlMD5", XML1_HASH))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$[0].name").value("sip"))
                 .andExpect(jsonPath("$[0].consistent").value(true))
@@ -148,11 +145,9 @@ public class AipApiTest extends DbTest implements ApiTest {
     public void updateXmlTest() throws Exception {
         MockMultipartFile xmlFile = new MockMultipartFile(
                 "xml", "xml", "text/plain", XML2_ID.getBytes());
-        MockMultipartFile metaFile = new MockMultipartFile(
-                "meta", "meta", "text/plain", XML2_HASH.getBytes());
 
         mvc(api)
-                .perform(MockMvcRequestBuilders.fileUpload(BASE + "/{sipId}/update", SIP_ID).file(xmlFile).file(metaFile))
+                .perform(MockMvcRequestBuilders.fileUpload(BASE + "/{sipId}/update", SIP_ID).file(xmlFile).param("xmlMD5", XML2_HASH))
                 .andExpect(status().isOk())
                 .andExpect(jsonPath("$.name").value("xml"))
                 .andExpect(jsonPath("$.consistent").value(true));
